agents/sk_orchestrator/requirements.txt
semantic-kernel>=1.15.0
azure-functions>=1.21.0
azure-identity>=1.19.0
